# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cash Bank apis",
        "contact": {
            "name": "Yosof Bayan",
            "url": "https://wa.me/+963967213544",
            "email": "yosofbayan75@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "local Base URL"
        },
        {
            "url": "https://cashbank-back.icrcompany.net/api/v1",
            "description": "Develop Base URL"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "App",
                    "App - Auth"
                ],
                "summary": "register a new user",
                "description": "register a new user with the provided information",
                "operationId": "app/register",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "username",
                                    "phone_number",
                                    "email",
                                    "password",
                                    "verification_code"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "referral_code": {
                                        "type": "string"
                                    },
                                    "verification_code": {
                                        "type": "string",
                                        "example": "0123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}